refactors project structure
