+++ src/model/configuration.ts
