+++ src/model/responseModel.ts
