+++ src/model/typedRequest.ts
